Relocate dns tests to test_dns (#557)
